Agregando End Points a los Controller

+++ SistemaDeVentasCoder/Models/Producto.cs
@@ -4,8 +4,8 @@
     {
         public int Id { get; set; }
         public string Descripciones { get; set; }
-        public decimal Costo { get; set; }
-        public decimal PrecioVenta { get; set; }
+        public double Costo { get; set; }
+        public double PrecioVenta { get; set; }
         public int Stock { get; set; }
         public int IdUsuario { get; set; }
 
@@ -18,7 +18,7 @@
             Stock = 0;
             IdUsuario = 0;
         }
-        public Producto(int id, string descripciones, decimal costo, decimal precioVenta, int stock, int idUsuario)
+        public Producto(int id, string descripciones, double costo, double precioVenta, int stock, int idUsuario)
         {
             Id = id;
             Descripciones = descripciones;
